Added proper support for TSV final output Added proper support for converting to TSV for rename Fixed an issue with Subfile repr

+++ src/lib/tools/bigFileWriter.py
@@ -4,6 +4,8 @@ import lib.commonFuncs as cmn
 import pandas as pd
 import sys
 from enum import Enum
+import pyarrow as pa
+import pyarrow.parquet as pq
 
 class Format(Enum):
     CSV = "csv"
@@ -23,7 +25,7 @@ class Subfile:
         self.filePath = location / f"{fileName}.{Format(format).value}"
 
     def __repr__(self) -> str:
-        return self.filePath
+        return str(self.filePath)
     
     @classmethod
     def fromFilePath(cls, filePath: Path) -> 'Subfile':
@@ -38,19 +40,20 @@ class Subfile:
     def read(self) -> pd.DataFrame:
         return pd.read_csv(self.fullPath)
 
-    def rename(self, newFilePath: Path) -> None:
-        newFileFormat = Format(newFilePath.suffix)
-
+    def rename(self, newFilePath: Path, newFileFormat: Format) -> None:
         if newFileFormat == self.fileFormat:
             self.filePath.rename(newFilePath)
             return
         
-        if self.newFileFormat == Format.CSV:
-            df = self.read()
+        df = self.read()
+        
+        if newFileFormat == Format.CSV:
             df.to_csv(newFilePath, index=False)
+
+        elif newFileFormat == Format.TSV:
+            df.to_csv(newFilePath, sep="\t", index=False)
             
-        elif self.newFileFormat == Format.PARQUET:
-            df = self.read()
+        elif newFileFormat == Format.PARQUET:
             df.to_parquet(newFilePath, "pyarrow", index=False)
 
         self.remove()
@@ -81,6 +84,7 @@ class PARQUETSubfile(Subfile):
 class BigFileWriter:
     def __init__(self, outputFile: Path, subDirName: str = "chunks", sectionPrefix: str = "chunk", subfileType: Format = Format.PARQUET) -> 'BigFileWriter':
         self.outputFile = outputFile
+        self.outputFileType = Format(outputFile.suffix[1:])
         self.subfileDir = outputFile.parent / subDirName
         self.sectionPrefix = sectionPrefix
         self.subfileType = subfileType
@@ -113,7 +117,7 @@ class BigFileWriter:
         self.writtenFiles.append(subfile)
         self.globalColumns = cmn.extendUnique(self.globalColumns, df.columns)
 
-    def oneFile(self) -> None:
+    def oneFile(self, removeOld: bool = True) -> None:
         if self.outputFile.exists():
             print(f"Removing old file {self.outputFile}")
             self.outputFile.unlink()
@@ -121,13 +125,25 @@ class BigFileWriter:
         if len(self.writtenFiles) == 1:
             print(f"Only single subfile, moving {self.writtenFiles[0]} to {self.outputFile}")
 
-            self.writtenFiles[0].rename(self.outputFile)
+            self.writtenFiles[0].rename(self.outputFile, self.outputFileType)
             self.subfileDir.rmdir()
             return
 
         print("Combining into one file")
+        if self.outputFileType in (Format.CSV, Format.TSV):
+            self.oneCSV(removeOld)
+        elif self.outputFileType == Format.PARQUET:
+            self.oneParquet(removeOld)
+
+        print(f"\nCreated a single file at {self.outputFile}")
+        self.subfileDir.rmdir()
+        self.writtenFiles.clear()
+
+    def oneCSV(self, removeOld: bool = True):
+        delim = "\t" if self.outputFileType == Format.TSV else ","
+
         with open(self.outputFile, 'w', newline='', encoding='utf-8') as fp:
-            writer = csv.DictWriter(fp, self.globalColumns)
+            writer = csv.DictWriter(fp, self.globalColumns, delimiter=delim)
             writer.writeheader()
 
             fileCount = len(self.writtenFiles)
@@ -138,8 +154,14 @@ class BigFileWriter:
                 for row in df.to_dict(orient="records"):
                     writer.writerow(row)
 
-                file.remove()
+                if removeOld:
+                    file.remove()
+        
+    def oneParquet(self, removeOld: bool = True):
+        schema = pa.schema([(column, pa.string()) for column in self.globalColumns])
+        with pq.ParquetWriter(self.outputFile, schema=schema) as writer:
+            for file in self.writtenFiles:
+                writer.write_table(pq.read_table(str(file)))
 
-        print(f"\nCreated a single file at {self.outputFile}")
-        self.subfileDir.rmdir()
-        self.writtenFiles.clear()
+                if removeOld:
+                    file.remove()
